pyproject.toml
[project]
name = "aws-strands-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "strands-agents>=1.19.0",
    "strands-agents-tools>=0.2.17",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.0",
]
